Anton/wip migrations

+++ packages/backend/src/lib/models/factories/sequelize/sharedMethods/synchronizeFactory.js
@@ -1,9 +1,16 @@
-module.exports = function synchronizeFactory({ Model } = {}) {
+module.exports = function synchronizeFactory({ config, Model } = {}) {
+  if (!config) {
+    throw new Error('Have to provide config')
+  }
   if (!Model) {
     throw new Error('Have to provide model')
   }
 
   return function synchronize({ force = false } = {}) {
+    if (!config.db.allowSync) {
+      return Promise.resolve()
+    }
+
     return Model.sync({ force })
   }
 }
